Feature/signupdetail

See merge request s09-fintech-finance-sub2/S09P22B207!14

+++ Front-End/pro/app/src/main/java/com/example/project/AuctioncreatePage.kt
@@ -29,7 +29,7 @@ fun AuctionCreatePage(navController: NavHostController) {
     var currentPage by remember { mutableStateOf(1) }
     val itemsPerPage = 10
 
-    var selectedItemIndex by remember { mutableStateOf<Int?>(null) }
+    var selectedItemIndex by remember { mutableStateOf<Long?>(null) }
 
     Box(modifier = Modifier.fillMaxSize()) {
         Column(
