+++ Front-End/pro/app/src/main/java/com/example/project/api/LoginService.kt
@@ -15,7 +15,7 @@ interface LoginService {
 
 // 패턴 검증 요청에 대한 DTO
 data class PatternVerificationRequest(
-    val user_idx: Int,
+    val user_idx: Long,
     val pattern: String
 )
